+++ ChangeLog.md
@@ -4,6 +4,7 @@
 
 * Fix markup for 404 page
 * Allow trailing slash in paths
+* Redirect from / to /events/
 
 ## 1.0.0  -- 2017-11-06